1. Изменил вывод ссылок на сайте с использованием WordPress диррективы site_url(). 2. Добавил выбор страниц при переключении списка с пользователями.    2.1. Реализовал асинхронный (Ajax) вывод списка пользователей.    2.2. Реализовал асинхронное (Ajax) переключение страниц списка пользователей.

+++ plugins/askue/pages/meters_table.php
@@ -20,6 +20,8 @@ function showNestedObjects($nestedLevel, $energyObjectId, $dataController, $nest
     $nestedEnergyObjects = $dataController->selectNestedEnergyObjects($energyObjectId, $user_id);
     $nestedParentMeters = $dataController->selectMetersList($energyObjectId);
 
+    //echo "user_id: ".$user_id." | count(nested_objects)=".count($nestedEnergyObjects);
+
     $parentEnergyObject = $dataController->selectEnergyObject($energyObjectId);
     if(!empty($parentEnergyObject->getMeterId())) {
         for($i = 0; $i <count($nestedParentMeters); $i++) {
@@ -67,6 +69,10 @@ function printEnergyObjectRow($energyObject, $nestedLevel, $is_last_row, $parent
     //$rand_num = number_format (30000/rand(20, 50), 3);
     $energy_object_value = 0;
 
+    $energy_object_edit_url_admin = site_url('/wp-admin/admin.php?page=add_energy_object&edit='.$energyObject->getId());
+    $energy_object_edit_url_user_room = site_url('/user-room/meters-management/add-energy-object?edit='.$energyObject->getId());
+
+
     if($object_value != null) $energy_object_value = $object_value;
 
     $padding_left = 10 + 10*$nestedLevel;
@@ -94,14 +100,14 @@ function printEnergyObjectRow($energyObject, $nestedLevel, $is_last_row, $parent
                     echo '<div class="nested_name"><div class="nested_name_text">'.$energyObject->getName().'</div></div></div></td>
                     <td></td>
                     <td style="text-align:right; padding-right:10px;">'.$energy_object_value.'</td>';
-    //if($access_level == 2 || $access_level == 3) {
+    if(ACCESS_LEVEL == 2 || ACCESS_LEVEL == 3) {
         echo '         <td>
                         <div style="padding-top:6px;">
                             <div style="display: inline-block;">';
                                 if(is_admin())
-                                    echo '<div class="edit-button" onclick="edit_button_clicked = true; location.href=\'/wp-admin/admin.php?page=add_energy_object&edit='.$energyObject->getId().';\'"></div>';
+                                    echo '<div class="edit-button" onclick="edit_button_clicked = true; location.href=\''.$energy_object_edit_url_admin.';\'"></div>';
                                 else
-                                    echo '<div class="edit-button" onclick="edit_button_clicked = true; location.href=\'/user-room/meters-management/add-energy-object?edit='.$energyObject->getId().';\'"></div>';
+                                    echo '<div class="edit-button" onclick="edit_button_clicked = true; location.href=\''.$energy_object_edit_url_user_room.';\'"></div>';
                             echo '</div>
                             <div style="display: inline-block; padding-left:5px;">
                                 <div class="delete-button" onclick="edit_button_clicked = true; delete_energy_object('.$energyObject->getId().',\''. $energyObject->getName().'\')"></div>
@@ -109,7 +115,7 @@ function printEnergyObjectRow($energyObject, $nestedLevel, $is_last_row, $parent
                         </div>
                     </td>
                 </tr>';
-    //}
+    }
 }
 
 function printMeterRow($meter, $nestedLevel, $is_last_row, $last_value, $parent_meters_count) {
@@ -117,15 +123,18 @@ function printMeterRow($meter, $nestedLevel, $is_last_row, $last_value, $parent_
     $last_value_val = 0;
     if($last_value != null) $last_value_val = $last_value->getValue();
 
+    $meter_edit_url_admin = site_url('/wp-admin/admin.php?page=add_meter&edit='.$meter->getId());
+    $meter_edit_url_user_room = site_url('/user-room/meters-management/add-meter?edit='.$meter->getId());
+
     $padding_left = 10+10*$nestedLevel;
 
     if($nestedLevel > 1) $padding_left = 20;
 
     if(is_admin()) {
-        $url_meter_detail = '/wp-admin/admin.php?page=meter_details&meter='.$meter->getId();
+        $url_meter_detail = site_url('/wp-admin/admin.php?page=meter_details&meter='.$meter->getId());
     }
     else {
-        $url_meter_detail = '/user-room/meter-management/meter-details?meter='.$meter->getId();
+        $url_meter_detail = site_url('/user-room/meter-management/meter-details?meter='.$meter->getId());
     }
 
     //echo '<tr onclick="window.location.href=\''.$url_meter_detail.'\';">
@@ -151,14 +160,14 @@ function printMeterRow($meter, $nestedLevel, $is_last_row, $last_value, $parent_
                     echo '<div class="nested_name"><div class="nested_name_text">'.$meter->getName().'</div></div></div></td>
                     <td>'.$meter->getNum().'</td>
                     <td style="text-align:right; padding-right:10px;">'.$last_value_val.'</td>';
-    //if($access_level == 2 || $access_level == 3) {
+    if(ACCESS_LEVEL == 2 || ACCESS_LEVEL == 3) {
     echo '         <td>
                         <div style="padding-top:6px;">
                             <div style="display: inline-block;">';
                                 if(is_admin())
-                                    echo '<div class="edit-button" onclick="edit_button_clicked = true; location.href=\'/wp-admin/admin.php?page=add_meter&edit='.$meter->getId().';\'"></div>';
+                                    echo '<div class="edit-button" onclick="edit_button_clicked = true; location.href=\''.$meter_edit_url_admin.';\'"></div>';
                                 else
-                                    echo '<div class="edit-button" onclick="edit_button_clicked = true; location.href=\'/user-room/meters-management/add-meter?edit='.$meter->getId().';\'"></div>';
+                                    echo '<div class="edit-button" onclick="edit_button_clicked = true; location.href=\''.$meter_edit_url_user_room.';\'"></div>';
 
                             echo '</div>
                             <div style="display: inline-block; padding-left:5px;">
@@ -167,32 +176,39 @@ function printMeterRow($meter, $nestedLevel, $is_last_row, $last_value, $parent_
                         </div>
                     </td>
                 </tr>';
-    //}
+    }
 }
 
 ?>
 
-<?php foreach ($energyObjectsList as $energyObject): ?>
+<?php //echo "access_level = ".ACCESS_LEVEL; ?>
 
+<?php foreach ($energyObjectsList as $energyObject): ?>
     <?php
     $customer = $dataController->selectCustomer($energyObject->getCustomerId());
+
+    $energy_object_edit_url_admin = site_url('/wp-admin/admin.php?page=add_energy_object&edit='.$energyObject->getId());
+    $energy_object_edit_url_user_room = site_url('/user-room/meters-management/add-energy-object?edit='.$energyObject->getId());
+    $energy_object_value = $dataController->selectEnergyObjectValue($energyObject);
+
     ?>
-    <table class="energy-object-top-table">
+    <table class="energy-object-top-table" border="0">
         <tr>
-            <td width="90%">
+            <td>
                 <div class="energy-object-title"><?=$energyObject->getName();?></div>
                 <div style="display:inline-block;">
                     [<a href=""><?=$customer->getName();?></a>]
                 </div>
             </td>
-            <?php if($access_level == 2 || $access_level == 3): ?>
+            <td align="right" style="padding-right:10px;">Сумма <?=$energy_object_value;?> Кв/ч</td>
+            <?php if(ACCESS_LEVEL == 2 || ACCESS_LEVEL == 3): ?>
             <td align="center" width="6%">
                 <div style="padding-top:6px;">
                     <div style="display: inline-block;">
                         <?php if(is_admin()):?>
-                            <div class="edit-object-button" onclick="location.href='/wp-admin/admin.php?page=add_energy_object&edit=<?=$energyObject->getId();?>'"></div>
+                            <div class="edit-object-button" onclick="location.href='<?=$energy_object_edit_url_admin?>'"></div>
                         <?php else: ?>
-                            <div class="edit-object-button" onclick="location.href='/user-room/meters-management/add-energy-object?edit=<?=$energyObject->getId();?>'"></div>
+                            <div class="edit-object-button" onclick="location.href='<?=$energy_object_edit_url_user_room?>'"></div>
                         <?php endif; ?>
                     </div>
                     <div style="display: inline-block; padding-left:5px;">
@@ -211,16 +227,20 @@ function printMeterRow($meter, $nestedLevel, $is_last_row, $last_value, $parent_
                     <th width="33%" style="text-align:left; padding-left:10px;">Название</th>
                     <th>Счетчик №</th>
                     <th width="33%" style="text-align:right; padding-right:10px;">Потребление (Кв/ч)</th>
-                    <?php if($access_level == 2 || $access_level == 3): ?>
+                    <?php if(ACCESS_LEVEL == 2 || ACCESS_LEVEL == 3): ?>
                         <th width="6%"></th>
                     <?php endif; ?>
             </tr>
 
             <?php
-            if($access_level==3)
+            if(ACCESS_LEVEL==3) {
+                //echo "show nested obj 1";
                 showNestedObjects(0, $energyObject->getId(), $dataController);
-            else
-                showNestedObjects(0, $energyObject->getId(), $dataController, $customer->getId());
+            }
+            else {
+                //echo "show nested obj 2<br>".print_r($customer);
+                showNestedObjects(0, $energyObject->getId(), $dataController, array(), $customer->getId());
+            }
             ?>
         </table>
